+++ BlasII.QualityOfLife/QolSettings.cs
@@ -7,6 +7,11 @@ namespace BlasII.QualityOfLife;
 public class QolSettings
 {
     /// <summary>
+    /// Should cutscenes and quotes be skipped
+    /// </summary>
+    public bool CutsceneSkip { get; set; }
+
+    /// <summary>
     /// Should the typhoon attack be made fps-independent
     /// </summary>
     public bool ConsistentTyphoon { get; set; }
