+++ BlasII.QualityOfLife/StorySkip.cs
@@ -38,39 +38,7 @@ internal class Tutorial_Skip2_Patch
     }
 }
 
-/// <summary>
-/// Level 3 - Skip cutscenes
-/// </summary>
-[HarmonyPatch(typeof(PlayCutscene), nameof(PlayCutscene.OnEnter))]
-class Cutscene_Skip_Patch
-{
-    public static bool Prefix(PlayCutscene __instance)
-    {
-        if (Main.QualityOfLife.CurrentSettings.SkipStoryLevel < 3)
-            return true;
 
-        // Don't skip Eviterno defeat cutscene
-        if (__instance.cutsceneId?.name == "CTS17_id")
-            return true;
-
-        ModLog.Warn("Skipping cutscene: " + __instance.cutsceneId?.name);
-        __instance.Finish();
-        return false;
-    }
-}
-[HarmonyPatch(typeof(ShowQuote), nameof(ShowQuote.OnEnter))]
-class Quote_Skip_Patch
-{
-    public static bool Prefix(ShowQuote __instance)
-    {
-        if (Main.QualityOfLife.CurrentSettings.SkipStoryLevel < 3)
-            return true;
-
-        ModLog.Warn("Skipping quote: " + __instance.Owner.name);
-        __instance.Finish();
-        return false;
-    }
-}
 
 /// <summary>
 /// Level 4 - Skip story events
